Debug the run. There have prblem in emus, If enum are empty it create a issue in foreginkey foremat. I fixed it. I also change the type of id, long to int

+++ backend/src/main/java/com/f/backend/controller/AuthController.java
@@ -49,7 +49,7 @@ public class AuthController {
     }
 
     @GetMapping("/active/{id}")
-    public ResponseEntity<String> activeUser(@PathVariable("id") long id) {
+    public ResponseEntity<String> activeUser(@PathVariable("id") int id) {
         try {
             String response = authService.activeUser(id);
             return ResponseEntity.ok(response);
